+++ Sources/PulseUI/Helpers/UXKit.swift
@@ -15,23 +15,23 @@ import SwiftUI
 
 struct Palette {
 #if os(watchOS)
-    static var red: UXColor { UXColor(Color.red) }
-    static var pink: UXColor { UXColor(Color.pink) }
+    static var red: UXColor { Palette.darkRed }
+    static var pink: UXColor { Palette.darkPink }
 #else
     static var red: UXColor {
         UXColor.dynamic(light: Palette.lightRed, dark: Palette.darkRed)
     }
-    
-    private static let lightRed = UXColor(red: 196.0/255.0, green: 26.0/255.0, blue: 22.0/255.0, alpha: 1.0)
-    private static let darkRed = UXColor(red: 254.0/255.0, green: 129.0/255.0, blue: 112.0/255.0, alpha: 1.0)
 
     static var pink: UXColor {
         UXColor.dynamic(light: Palette.lightPink, dark: Palette.darkPink)
     }
-    
+#endif
+
+    private static let lightRed = UXColor(red: 196.0/255.0, green: 26.0/255.0, blue: 22.0/255.0, alpha: 1.0)
+    private static let darkRed = UXColor(red: 254.0/255.0, green: 129.0/255.0, blue: 112.0/255.0, alpha: 1.0)
+
     private static let lightPink = UXColor(red: 155.0/255.0, green: 35.0/255.0, blue: 147.00/255.0, alpha: 1.0)
     private static let darkPink = UXColor(red: 252.0/255.0, green: 95.0/255.0, blue: 163.0/255.0, alpha: 1.0)
-#endif
 }
 
 #if os(macOS)